test: fix entry points in e2e
