+++ packages/playground/server-components/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/entry-client.jsx"></script>
+    <script type="module" src="@shopify/hydrogen/entry-client"></script>
   </body>
 </html>
