--- packages/playground/server-components/src/entry-client.jsx
@@ -1,7 +1,0 @@
-import renderHydrogen from '@shopify/hydrogen/entry-client';
-
-function ClientApp({children}) {
-  return children;
-}
-
-export default renderHydrogen(ClientApp, {});
